Update patch set 8
Patch Set 8:

I also played with the one-click editing use case in Impress (on staging) and sometimes the blinking cursor is there, but when I type, nothing happens. So the blinking cursor is displayed, but the actual focus is not on the shape. So no idea on what it depends whether these issues are coming up or not, but I can imagine that Jenkins hits that race condition which ruins the state of the document during the test.

Patch-set: 8