Update patch set 3
Patch Set 3: Code-Review+2

> Patch Set 3:
> given that all the places always pass the full length of the buffer, either just pass a std::string, or just a std::string_view if you want flexibility?

Session::sendTextFrame() didn't pass the full length of the buffer, as far as I see (and that buffer is not even null-terminated). So the 2 options I have is to either copy the sub-string to an std::string at one place (smaller patch, bug pointless copying) or change the interface to never require an std::string. I go with the later for now, we can always re-visit it if needed.

Patch-set: 3
Reviewer: Gerrit User 1000013 <1000013@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=+2