Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hmm, cypress test is still failing. Let's just ignore that for now.
> I disabled the failing test, later I'll check what is the issue there:
> https://cgit.freedesktop.org/libreoffice/online/commit/?id=4c57f43d60389b0368afc02394ff68411abeb49e

By looking at the failure (for the first time ever about cypress), it might be not expecting the new behaviour (jumping to the cursor). "failed because the center of this element is hidden from view"

Thank you for taking care of that!

Patch-set: 2
Reviewer: Gerrit User 1001264 <1001264@6622c302-437a-49ad-a91d-55ba7dcf1615>