Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Hmm ? surely we get cursor messages ~all the time - and we also sometimes want to scroll / pan to areas where the cursor is not visible: then again, I guess we don't get cursor messages then (?)
> 

We should, but we might (on zoom change for example). But this change shouldn't do anything if there is no real cursor change, and the previous cursor is not an empty rectangle.

> Presumably you've tested it left & right with panning away from the cursor, and selecting / manipulating shapes / charts etc. ?
> 
> Thanks !

Did some manual testing by panning away, and inserting shapes. Didn't see any obvious problems.

Patch-set: 4