Add some UNO commands that bring up tunneled dialogs to the mobileBlackList
Will do this in batches. These are the menu items that showed up in
the context menu for the "internal" chart in a chart object (when you
have selected the chart and "opened" it, so that the rest of the sheet
is darkened). (Unclear about terminology here, the terms in quotes are
ad-hoc.)

Change-Id: I138f992e1b9cc0b4e31a2c0b64bea2e9c160ce4f