leaflet: maintain the keyboard state after toolbar actions
Previously we supressed the keyboard after toolbar
actions, but that is problematic because then the
user is unable to delete/replace selected text,
because the keyboard is hidden and there is no way
of showing keyboard without changing the selection.

Now we maintain the keyboard state, which is likely
visible, since a selection shows the keyboard.

This might not be ideal, because the user might
hide the keyboard on the device and we will restore
it after invoking a toolbar action, but at least
it's more usable now. Unfortunately, there is no
API to track the keyboard visibility.

New Cypress tests added to validate the above.

Change-Id: If0f2e96909ff20753043734789397d190cedb502