Update patch set 10
Patch Set 10:

> You also expose map directly to the cypress test framework. One step forward and you don't interact with the UI but the internal JS objects.

See also another patch, where it already happened:
https://gerrit.libreoffice.org/c/online/+/90524

This test framework's aim is to simulate user interaction as closely as possible. If you use internal JS objects and methods, then we will loose that. It's easy to use internal JS methods, so if you introduce that as a practice it will spread in the code and so the test will get far from the normal user interaction.

Patch-set: 10