Update patch set 4
Patch Set 4:

Thanks ! after more reading, I'm personally itching to fix this one - it looks horribly misleading. Then again - the behavior of mobile/popup keyboards is extremely weird and (I guess) the problem is that we can end up in a state where the browser thinks there is a valid, focused input - but yet somehow we've managed to now show the on-screen keyboard (and this is a 'feature') - so we can't check that by asking the browser. But if it is to track this horror - then it needs a deeply explicit name for the method (and -really- good documentation explaining what it is about) - and this patch is missing that.

Patch-set: 4
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>