wsd: improve lifecycle in ClientSession::onDisconnect()
Once unit-bad-doc-load completes (with success in exitTest()), sometimes
we have an error during shutdown.

The reason seems to be that ClientSession::onDisconnect() calls
DocumentBroker::removeSession(), which may delete the ClientSession, so
by time time isCloseFrame() is called, we have trouble.

Fix the problem by keeping a reference to self before calling
removeSession().

Change-Id: If5b409822563ba5a45d453329516671065d8f054