cypress: mobile: restore original selectAllMobile() in writer_helper.js.
Using CTRL+A is simpler, but with using the Hamburger menu we can
make sure that menu item is working. CTRL+A is not an actual
feature on mobile. Also it's good that we use hamburger menu, so
we can check interaction of hamburger menu and other wizards
using the same bottom panel.

Change-Id: I640e515070046ad18b00be69509ed895844962a0