leaflet: do not expose map to cypress tests
We only need to expose a couple of helpers
to test the keyboard and clipboard, which
are not possible to test via the UI.

This patch reduces the surface area of
exposed code to testing, to avoid misuse.
The only needed helpers should be made
available, and they are read-only functions,
so there is no risk of changing the internal
state from test code.

Change-Id: Id4a1800e3d3b9364754a18fb11f61df612e56490