Add two pseudo-assertions for code that IMHO will never be executed
I suggest we introduce this convention for such assertions.
Feel free to modify the message syntax. The idea is that if you notice
some code block you think will never be executed, add a comment and a
console.log with a message in this syntax. If somebody then later sees
that logging appear after all, they can add/modify a comment describing
what is actually going on.

Change-Id: I76302e414d78912b6d20ba149c69d00d995613c6