Add two pseudo-assertions for code that IMHO will never be executed
I suggest we try to introduce this convention for such assertions.
Feel free to modify the message syntax. The idea is that if you notice
some code block that should never be executed, add a console.log with
this syntax. If somebody then later sees that logging appear, they can
modify a comment describing what is actually going on.

Change-Id: I76302e414d78912b6d20ba149c69d00d995613c6