Get rid of the iOS app specific "context toolbar"
The less special-case code/logic/UI, the better the world is.

It had weird unintended side effects anyway. (In the iOS app, which
was the only place where it was used.) For instance, in a spreadsheet
document on iPhone, after loading the document, the normal "mobile
wizard" style context menu did not appear, no matter how long you
pressed a selected cell. But if you first got the context toolbar to
show up once, and after that selected some cell and long-pressed it,
the mobile wizard style context menu did appear, just as when using
web-based Online on the same device.

Note that despite that the code in Control.ContextToolbar.js was used
only in the iOS app (L.control.contextToolbar() was called only if
window.ThisIsTheiOSApp), the file did contain code for non-iOS, too.

Change-Id: Icdbfc7e71f77539c1bcb5ac8cf6984375297a4bd