mobile: don't show a dummy mobile wizard before sidebar content arrives.
If we show the mobilw wizard instantly after the mobile
wizard toolbar button is pushed, then mobile wizard will
be an empty bottom panel with an undefined content. It's
better to show the content, only after we get the
sidebar content from core.
For example, if the hamburger menu was displayed in the
bottom panel previously then we will see a white screen
first when we push the mobile wizard toolbar button. This
white screen is the background layout of the mobile wizard.

It looks wierd and also makes testing hard since we have
some wierd dummy HTML structures.

Change-Id: I852bce1ca1460b2210229bec294a249a47523099