Avoid 'TypeError: undefined is not an object' in the iOS (and Android?) app
There is no map._clip in the mobile apps as they don't use the
internal clipboard implementation in loleaflet.

Not sure why L.Clipboard.filterExecCopyPaste() needs a check for
window.ThisIsAMobileApp as no L.Clipboard object is ever created in
the mobile apps. (See L.Map.initialize() where this._clip is not
assinged any value in the iOS or Android app cases.)

Change-Id: Ib27ec10463667091e65fd813ceb554e83c0952fb