leaflet: refactor _launchCalcInputBar()
Preparing to avoid re-creating the input-bar
on every update message. This will not just avoid
flicker and unnecessary CPU utilization, but
it will make Cypress tests more reliable, since
destroying an HTML element can fail tests, even
if they are created right after (it becomes a
matter of random timing whether a test that checks
the presence of such an element passes or fails).

Change-Id: I8ff887578ebd719b05bb62059d3f0a0bbb929898