mobile: introduce mobile_mode class for body element.
We can decide whether we are on mobile (phone) in the JS
code and with adding a 'mobile_mode' class to the body
element we can use this information in the css files too:
e.g. body.mobile_mode #toolbar-hamburger css selector.

The problem is that we have more inconsistent media rules
in css files for mobile phones:
@media (max-width: 767px), (max-device-height: 767px)
@media (max-width: 767px) and (orientation: portrait), (max-width: 1023px) and (orientation: landscape)
@media (max-width: 767px), (max-device-height: 767px) and (pointer: coarse)

These are all for mobile phone detection. It's not enough that
these rules are inconsistent, but JS code also has a mobile
phone detection, which also different. So instead of changing
media rules back and there and also breaking mobile emulation
with that, a better solution to make the decision at one place
in the JS code and use that decision everywhere.

The problem of detecting whether we are on a mobile phone
is still an issue, but probably we can solve it easier
in the JS code, then in the CSS.

Also break rules with multiple selectors to
multiple lines. Somewhere we had ~1000 characters long
line in the css file.

Change-Id: I012433f85df1b64c1ab2de2b6f8612e858a78a45