Update patch set 10
Patch Set 10:

Pushed to master by accident - I stripped out the Android stuff - can you turn that into a new patch (it was making CI fail).

Also - I killed the General.hpp - and moved that into Util.cpp - the #pragma once is also novel - best not to use new constructs that are not used elsewhere in the code.

So I did a few things like that and got it in =) when you git pull -r - do a git rebase --skip # and you should be fine.

Thanks !

Patch-set: 10