leaflet: support pasting in the search box
When pasting into the document a special handler
is used to manage the transfer of data. For the search
box we need to let the default handler do the right thing
and so we check that we aren't searching before we override
the paste handler logic, which explicitly pastes into the doc.

Change-Id: I570168d1aaf70c2a78403644e006ba9625ec68db