Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> This looks very nice - thanks ! =) CI having some problems here; but does 'make check' pass for you.

'make check' passed on my local build.

> Good to see lots of things cleaned up there, and some homogeneity arriving. I forget what we wanted to pass through this websocket to make things more 
> beautiful, but ... it looks like a nice cleanup anyway.

We need to pass the count of segmentation fault errors occurred in kit processes from ForKit to LOOLWSD admin and thus we need communication in both ways. After this patch will be accepted I will make a new patch based on this one for 'segmentation fault' metric.

> IIRC we were going to pass file-descriptors over the UDS back to the Admin client for the /proc/smaps for each process (or did I go a bit loopy ;-) - and/or was that another problem.

This is a different problem related to memory usage reporting from kit processes. A kit process is not able to report memory usage to LOOLWSD if it hangs or is just in the middle of heavy processing (a kit has only one thread). And we cannot rely on this reporting for identifying problematic kits which use too much memory (in this case they usually hang). This is needed to avoid restarting an entire pod in Kubernetes (because the memory limit was reached) by killing exactly this kind of kit processes.

Also, I think a review from "mobile" perspective is needed for this patch just to avoid issues similar to those from my previous patch.

> Thanks !

Thank you too !

Patch-set: 1
Reviewer: Gerrit User 1002093 <1002093@6622c302-437a-49ad-a91d-55ba7dcf1615>