JWTAuth::verify: avoid working with RTTI without a reason
The purpose of the convert() function (without arguments) is to either
return fast in case the type of the JSON value and the argument match or
do a conversion.

We generate this JSON, so we know the JSON type will be always a string
and the argument type will be a size_t, so use the variant that always
does a conversion.

(Additional benefit is that I seem to hit a false alarm with
admin_fuzzer in the old code and it goes away with this change.)

Change-Id: I40851ab3ddd46fb1515ea9f0c9f40d9ec2006de7