On iOS make the document show up as modified also after auto or explicit save
It is not enough to just have the core code rewrite the document file.
We need to do the same dance as in -[DocumentViewController bye].

I wonder if it would be cleaner to copy the document file to some
temporary location when opening it, and use only that copy in Online
and core, and then after core has saved it (either for an auto-save,
an explicit save, or when closing the document) do the -[UIDocument
saveToURL:forSaveOperation:completionHandler] thing?

Change-Id: I78fec08a90424780dc6e794347a0acf2c64e1dab