Update patch set 2
Patch Set 2: Code-Review-1

Thanks - so; there are two help->about dialogs in Android.
I would guess we want to have the OS version information in the normal hamurger -> about dialog that you get when editing a file.

The patch I tested added just a '9' in the license dialog; which if we go this way, would want a "Android " prefix to it =)

The Hamburger->about dialog still had a %OS_VERSION% in it - so ...

Patch-set: 2
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1