Fix problems after my 293f4913d2cdfe5385e2cdc0e3bebde281da1578
It's enough to call the -[UIDocument
saveToURL:forSaveOperation:completionHandler:] only in
DocumentBroker::sendUnoSave(). And on the other hand, in
-[DocumentViewController bye] we can't want for the
LOOLWSD::lokit_main_mutex as the main queue is needed for parts of
what the saveToURL does.

There is still some problem, though. Occasionally when closing the
document the app hangs.

Change-Id: Idb65be23a7cb6ad1288fbbd23c7471e0fb8d52f4