mobile: Never use L.Browser.mobile, instead use window.mode.isXYZ().
The L.Browser detects the browser capabilities, but the
window.mode.isMobile()/isTablet()/isDesktop() is what decides what form
factor are we targeting in the particular instance of the Online.

This commit tries to kill convert the L.Browser.mobile checks into the
appropriate window.mode.isXYZ(), hope I got the intention right on most
of the places - it wouldn't be correct to blindly convert everything to
(isMobile() || isTablet()).

Change-Id: I57430058ec15deb5f67f9c645d234f998fbf539d