cypress: mobile: improve copying via context menu.
So we don't need to close the mobile wizard before copying.
The main issue was with contains() method, which first
selects the HTML items and waits until these items gets
the content 'Copy'. What we actually need is to wait until
an item appears with the content 'Copy'. Not the existing
items's content is updated with the 'Copy' text, but new
items are created.

Change-Id: I235d2756ce41a099267fa335c215d3eca6f468c2