Update patch set 1
Patch Set 1:

Andras has a phone that has:

window.screen.width:  869
window.screen.height: 412

when the UI is started in the landscape mode.  But it should be still treated as a mobile phone, not a tablet, that's why the check for the height too.

I think it is not a bad thing when small tablets will actually get the phone UI too; but obviously I don't want to break iPad here :-)

Patch-set: 1
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>