cypress: introduce selectItemByContent() method instead of contains().
cy.get().contains() behaves unexpectedly. With pseudo code:

// First find the item using the selector
do {
  item = get(selector)
} while (item == undefined)

// Then wait until it get's the right content.
// Item points to an object which can change it time.
do {
} while (item.content != specified_content)


What we actually need is something like this:
// Wait for an item which matches both the selector and content.
do {
  item = get(selector)
} while (item == undefined || item.content != specified_content)

The new selectItemByContent() method behaves like this.

It just looks wierd when you experience a test failure in the
interactive test runner and you check the DOM and you can see
the right item with the rigth content and the test still fails.

Change-Id: I0b1466a165451e831401e1f4b500ce16bd3f38b0
Signed-off-by: Tamás Zolnai <tamas.zolnai@collabora.com>