Update patch set 7
Patch Set 7:

Sure - it's no problem; Gokay can unwind this for your change I'm sure.

Wrt. showing OS version details - we do that for the server-side on Linux now, and yes the Ubuntu <NN> vs. openSUSE <VV> difference has been the cause of bugs - eg. this inode leakage, so really useful to have that detail.

Actually as a follow-on patch Gokay - it would be lovely to get the OS version details dumped in the start-up log message we shove out: we get a lot of logs out of people showing problems, and would be really nice to have that there.

I assume in future we will have things that fail to work on a given version of Android; and having that easily to hand would be really helpful.

Patch-set: 7
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>