wsd: faster jail setup via bind-mount
loolmount now works and supports mounting
and unmounting, plus improved logging.
When enabled, binding improves the jail
setup time by anywhere from 2x to orders
of magnitude.

A new config entry mount_jail_tree
controls whether mounting is used or
the default linking/copying of jail
contents. It is set to true by default
and falls back to linking/copying.
A test mount is done when the setting is
enabled, and if mounting fails, it's
disabled to avoid noise.

The network/system files in /etc that need
frequent refreshing are now updated
in systemplate to facilicate mounting
the etc directory directly. These files
can change during the course of loolwsd
lifetime, and are unlikely to be updated
after installation at all. We link to them
in the systemplate/etc directory, and
if that fails, we copy them before forking
each kit instance to have the latest.

Also, the random temp directory is now
created and set correctly, plus many
misc logging and other improvements.

Change-Id: Iae3fda5e876cf47d2cae6669a87b5b826a8748df