wsd: faster jail setup via bind-mount
loolmount now works and supports mounting and
unmounting, plus improved logging.  When enabled,
binding improves the jail setup time by anywhere
from 2x to orders of magnitude.

A new config entry mount_jail_tree controls
whether mounting is used or the default
linking/copying of jail contents. It is set to
true by default and falls back to linking/copying.
A test mount is done when the setting is enabled,
and if mounting fails, it's disabled to avoid noise.

The network/system files in /etc that need frequent
refreshing are now updated in systemplate to
facilicate mounting the etc directory directly.
These files can change during the course of loolwsd
lifetime, and are unlikely to be updated after
installation at all. We link to them in the
systemplate/etc directory, and if that fails, we
copy them before forking each kit instance
to have the latest.

This reworks the approach used to bind-mount the
jails and the templates such that the total is
now down to only three mounts: systemplate, lo, tmp.

As now systemplate and lotemplate are shared, they
must be mounted as readonly, this means that user
must now be moved into tmp/user/ which is writable.

The mount-points must be recursive, because we mount
lo/ within the mount-point of systemplate (which is
the root of the jail). But because we (re)bind
recursively, and because both systemplate and
lotemplate are mounted for each jails, we need to
make them unbindable, so they wouldn't multiply the
mount-points for each jails (an explosive growth!)

The random temp directory is now created and set
correctly, plus many logging and other improvements.

Change-Id: Iae3fda5e876cf47d2cae6669a87b5b826a8748df