wsd: track failure to store a document to retry again
When we save a document, we immediately upload it to storage.
The document is of course marked as unmodified after
saving, but if Storage fails to store it, we will
never try again until and unless the document gets
modified (or the user forces saving).

By tracking the result of the last store attempt we
retry again on the next autoSave, which is invoked
upon shutdown as well, making sure that we'll at least
retry storing the document to storage.

Change-Id: I24a37d8e5f80a41e8bbcf6730a137c06419ace24