Make the "X" closebutton work also in the mobile apps
In normal cases it will not show up, but due to what is arguably a
bug, it sometimes shows up anyway. For instance, when you use the app
in split view on a tablet. So make the button work also in a mobile
app by posting the same 'BYE' message as done for the "<" closemobile
button.

Change-Id: Ifdc687ab734fa5e092999d94bd19fe1572281af9