tdf#128502: Get rid of the global TerminationFlag
Replace it with a local flag in the Socket class, and getters and
setters in related classes. Also introduce a separate flag that is set
when receiving a termination signal.

Getting rid of such global variables is necessary in order to be able
to have several documents open in the same process for the mobile
apps.

Change-Id: I64f2e1b236751a87e800808115e733ff13c91032