cypress: mobile: use a better workflow for using aliases / variables.
Since we are in an asynchronous environment, we can't just use
variables locally that easy. Cypress has alias feature to
introduce global variables and access them. Since these are
global we should init them anytime before use.
So a stable workflow seems to be this one:
* Set alias to an initial value (e.g. 0)
* Wait until it's actually set (e.g. cy.get().should())
* Set the alias to the value you need in the test
* Wait until it's actually set
* Move on with your test and assert on something using this alias

So we need indicators also for this variable to be set, before we
use it. This will make it work a sequential way, how we need it
to work.

Change-Id: I047373aa884b1ac7bac419cfb1811534c0e95862