Update patch set 7
Patch Set 7:

@Michael:

> Hmm; surely TerminationFlag is a way to shut down all of loolwsd; I'm really not convinced it's a good way to manage the lifecycle of individual documents in a multi-document setup (surely) ?

> Can we not remove the use of TerminationFlag to shut-down an individual document in the mobile code instead ?

I was not sure what all TerminationFlag is used for; I had the impression that it is used (in web-based Online) in several processes, not just loolwsd, and during the "normal" working of the process tree, not just when terminating the whole shebang. I will have to experiment. But if its purpose indeed is only to handle the forceful shutdown of the whole thing, then it has been used wrongly in the mobile apps all the time. In that case, it should not be set ever in the mobile apps.

Patch-set: 7
Reviewer: Gerrit User 1000062 <1000062@6622c302-437a-49ad-a91d-55ba7dcf1615>