Update patch set 7
Patch Set 7:

> I wonder, is it possible that the meaning and use of TerminationFag
> has changed over time, and the MOBILEAPP code still uses it in the
> old-fashioned way? Or have I misunderstood it from the start?

Not sure =) AFAICS it is used for forcing shutdown - Ctrl-C once - ShutdownRequestFlag set, Ctrl-C again and we get more aggressive with the TerminationFlag.

It looks like the ForKit uses it for poorly disconnected sockets - ie. if we don't do a clean handshake nicely (eg. LOOLWSD crashing) - which may be your problem; and (I guess) mobile does something with it, though I'd hope it is not useful in mobile.

Patch-set: 7
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>