Welcome: Add the dismiss/accept button
Now that we have the value of window.enableWelcomeMessageButton
coming from the loolwsd.xml, we can optionally enable/disable
the dismiss/accept button.

Using vex.dialog.open instead of vex.open to be able to
have the 'buttons' option.

Because the vex-dialog-message class displays the message
all-bold (very ugly) on mobile, we change it to vex-content,
just after opening the dialog.

We also limit the max height of the dialog to avoid double
scroll-bars.

Change-Id: I0c4d1cba303da49d23342a74538c7f340c4bcf1b