Update patch set 7
Patch Set 7:

@Michael:

So, do I understand correctly, that currently, in normal Online, the actual use of TerminationFlag is to cause the processes to exit even more bluntly, on the second receipt of a signal, in handleTerminationSignal()? The first time it sets ShutdownRequestFlag, the second time TerminationFlag.

Also, is it so that the remaining calls to setTerminationFlag() in the code are leftovers that won't actually be called when the server processes run normally and are terminated cleanly? (But can they be called in one process in some theoretical pathological situation when some (other) process is hung or whatever? But will the server recover from a such situation anyway, or will it have to be torn down completely? And, if the forkit loolwsd process is hung, nothing works. If the forkit process is hung, nothing works either. If an individual kit process is hung, the parent will kill it, won't it?)

1) In ForKit.cpp, after LOG_INF("Setting TerminationFlag due to 'exit' command from parent.");
Is this ever reached? At least quickly testing it seems not.

2) In ForKit.cpp, after LOG_WRN("ForKit connection lost without exit arriving from wsd. Setting TerminationFlag");
This does not seem to be reached either.

3) In Kit.cpp, after LOG_INF("Setting TerminationFlag due to 'exit' command.");
This is MOBILEAPP only, and based on the old or misunderstood semantics of TerminationFlag.

4) In Kit.cpp, after LOG_WRN("Kit connection lost without exit arriving from wsd. Setting TerminationFlag");
This does not seem to be reached. Quickly testing, the kit processes seem to always go away earlier already with the "Document [...] has no more views, exiting bluntly."

5) In Kit.cpp, after LOG_INF("Last session discarded. Setting TerminationFlag");
I don't see this either in the log file.

6) In LOOLWSD.cpp, after LOG_INF("Setting TerminationFlag");
Only in the fuzzing case.

7) In LOOLWSD.cpp, after LOG_INF("Setting TerminationFlag to avoid deadlocking unittest.");
Only in unit test.

Patch-set: 7
Reviewer: Gerrit User 1000062 <1000062@6622c302-437a-49ad-a91d-55ba7dcf1615>