ruler: make sure touch and mouse events don't cause an error
We need to make sure we don't accidentally get into either a
touch or mouse event codepath, which would cause an error as the
events and cases differe a bit. For example touch event has no
'button' member, which means a touch 'panstart' event would be
ignored in every case, which would then cause an error in
'panmove' and 'panend' because some variable wouldn't be set.

Change-Id: I08e4850b735f753639f7001b736c2883cf2d683f