Welcome: Prevent a million cookies during development
By default, cookies are stored with full path of
the web app, and we have an extension made of the
hash of the build. So each new commit means a new
web app for the browser, causing many cookies stored,
and the welcome dialog being shown after each commit.

From now on, we will be storing our cookies to the
'/loleaflet' path.

Change-Id: I4cb25ab0a53c4743cb929a7f6c20ab019da29409